Split search.rs (556 lines) into search/ subdirectory
Split crates/ui/src/search.rs into focused modules under search/:
- mod.rs: thin facade with pub mod/pub use + SearchPlugin
- types.rs: SearchState, BuildingResult, CitizenResult, constants
- helpers.rs: name generation arrays, citizen_name(), zone_label(), education_label()
- systems.rs: search_keybind, update_search_results, search_panel_ui
- tests.rs: unit tests

All public items re-exported from mod.rs — no external import changes needed.
