+++ crates/ui/src/search/systems.rs
@@ -1,9 +1,4 @@
-//! UX-044: Search/Filter for Buildings and Citizens.
-//!
-//! Provides a search bar (toggled via Ctrl+F) to find buildings and citizens.
-//! Buildings can be searched by zone type, level, or status (abandoned, under construction).
-//! Citizens can be searched by name, age, or occupation (education level).
-//! Results are displayed in a scrollable list; clicking a result jumps the camera.
+//! Systems for the search/filter feature: keybind handling, result updates, and UI rendering.
 
 use bevy::prelude::*;
 use bevy_egui::{egui, EguiContexts};
@@ -11,203 +6,11 @@ use bevy_egui::{egui, EguiContexts};
 use rendering::camera::OrbitCamera;
 use simulation::abandonment::Abandoned;
 use simulation::buildings::{Building, UnderConstruction};
-use simulation::citizen::{Citizen, CitizenDetails, Gender, HomeLocation, Position, WorkLocation};
+use simulation::citizen::{Citizen, CitizenDetails, HomeLocation, Position, WorkLocation};
 use simulation::config::CELL_SIZE;
-use simulation::grid::ZoneType;
 
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
-/// Maximum number of results to display per category.
-const MAX_RESULTS: usize = 50;
-
-// ---------------------------------------------------------------------------
-// Resources
-// ---------------------------------------------------------------------------
-
-/// Tracks the search panel state.
-#[derive(Resource)]
-pub struct SearchState {
-    /// Whether the search panel is visible.
-    pub visible: bool,
-    /// The current search query text.
-    pub query: String,
-    /// Whether to search buildings.
-    pub search_buildings: bool,
-    /// Whether to search citizens.
-    pub search_citizens: bool,
-    /// Cached building results: (entity, zone_label, level, status, grid_x, grid_y).
-    pub building_results: Vec<BuildingResult>,
-    /// Cached citizen results: (entity, name, age, education_label, grid_x, grid_y).
-    pub citizen_results: Vec<CitizenResult>,
-    /// Whether results need to be refreshed.
-    pub dirty: bool,
-    /// Track the previous query to detect changes.
-    prev_query: String,
-    /// Whether the text field should request focus on the next frame.
-    request_focus: bool,
-}
-
-#[derive(Clone)]
-pub struct BuildingResult {
-    pub entity: Entity,
-    pub zone_label: String,
-    pub level: u8,
-    pub status: &'static str,
-    pub grid_x: usize,
-    pub grid_y: usize,
-}
-
-#[derive(Clone)]
-pub struct CitizenResult {
-    pub entity: Entity,
-    pub name: String,
-    pub age: u8,
-    pub education: &'static str,
-    pub grid_x: f32,
-    pub grid_y: f32,
-}
-
-impl Default for SearchState {
-    fn default() -> Self {
-        Self {
-            visible: false,
-            query: String::new(),
-            search_buildings: true,
-            search_citizens: true,
-            building_results: Vec::new(),
-            citizen_results: Vec::new(),
-            dirty: false,
-            prev_query: String::new(),
-            request_focus: false,
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Name generation (matching citizen_info.rs)
-// ---------------------------------------------------------------------------
-
-const FIRST_NAMES_M: &[&str] = &[
-    "James", "John", "Robert", "Michael", "David", "William", "Richard", "Joseph", "Thomas",
-    "Daniel", "Matthew", "Anthony", "Mark", "Steven", "Paul", "Andrew", "Joshua", "Kenneth",
-    "Kevin", "Brian", "George", "Timothy", "Ronald", "Edward", "Jason", "Jeffrey", "Ryan", "Jacob",
-    "Gary", "Nicholas", "Eric", "Jonathan",
-];
-
-const FIRST_NAMES_F: &[&str] = &[
-    "Mary",
-    "Patricia",
-    "Jennifer",
-    "Linda",
-    "Barbara",
-    "Elizabeth",
-    "Susan",
-    "Jessica",
-    "Sarah",
-    "Karen",
-    "Lisa",
-    "Nancy",
-    "Betty",
-    "Margaret",
-    "Sandra",
-    "Ashley",
-    "Dorothy",
-    "Kimberly",
-    "Emily",
-    "Donna",
-    "Michelle",
-    "Carol",
-    "Amanda",
-    "Melissa",
-    "Deborah",
-    "Stephanie",
-    "Rebecca",
-    "Sharon",
-    "Laura",
-    "Cynthia",
-    "Kathleen",
-    "Amy",
-];
-
-const LAST_NAMES: &[&str] = &[
-    "Smith",
-    "Johnson",
-    "Williams",
-    "Brown",
-    "Jones",
-    "Garcia",
-    "Miller",
-    "Davis",
-    "Rodriguez",
-    "Martinez",
-    "Hernandez",
-    "Lopez",
-    "Gonzalez",
-    "Wilson",
-    "Anderson",
-    "Thomas",
-    "Taylor",
-    "Moore",
-    "Jackson",
-    "Martin",
-    "Lee",
-    "Thompson",
-    "White",
-    "Harris",
-    "Clark",
-    "Lewis",
-    "Robinson",
-    "Walker",
-    "Young",
-    "Allen",
-    "King",
-    "Wright",
-    "Hill",
-];
-
-fn citizen_name(entity: Entity, gender: Gender) -> String {
-    let idx = entity.index() as usize;
-    let first = match gender {
-        Gender::Male => FIRST_NAMES_M[idx % FIRST_NAMES_M.len()],
-        Gender::Female => FIRST_NAMES_F[idx % FIRST_NAMES_F.len()],
-    };
-    let last = LAST_NAMES[(idx / 31) % LAST_NAMES.len()];
-    format!("{} {}", first, last)
-}
-
-// ---------------------------------------------------------------------------
-// Display helpers
-// ---------------------------------------------------------------------------
-
-fn zone_label(zone: ZoneType) -> &'static str {
-    match zone {
-        ZoneType::None => "None",
-        ZoneType::ResidentialLow => "Residential (Low)",
-        ZoneType::ResidentialMedium => "Residential (Med)",
-        ZoneType::ResidentialHigh => "Residential (High)",
-        ZoneType::CommercialLow => "Commercial (Low)",
-        ZoneType::CommercialHigh => "Commercial (High)",
-        ZoneType::Industrial => "Industrial",
-        ZoneType::Office => "Office",
-        ZoneType::MixedUse => "Mixed Use",
-    }
-}
-
-fn education_label(education: u8) -> &'static str {
-    match education {
-        0 => "None",
-        1 => "Elementary",
-        2 => "High School",
-        3 => "University",
-        _ => "Advanced",
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Systems
-// ---------------------------------------------------------------------------
+use super::helpers::{citizen_name, education_label, zone_label};
+use super::types::{BuildingResult, CitizenResult, SearchState, MAX_RESULTS};
 
 /// Toggle search panel visibility with Ctrl+F.
 pub fn search_keybind(
@@ -479,78 +282,5 @@ pub fn search_panel_ui(
         orbit.focus.z = wy;
         // Zoom in closer to show the target
         orbit.distance = orbit.distance.min(400.0);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Plugin
-// ---------------------------------------------------------------------------
-
-pub struct SearchPlugin;
-
-impl Plugin for SearchPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<SearchState>().add_systems(
-            Update,
-            (search_keybind, update_search_results, search_panel_ui).chain(),
-        );
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_zone_label_all_variants() {
-        assert_eq!(zone_label(ZoneType::None), "None");
-        assert_eq!(zone_label(ZoneType::ResidentialLow), "Residential (Low)");
-        assert_eq!(zone_label(ZoneType::ResidentialMedium), "Residential (Med)");
-        assert_eq!(zone_label(ZoneType::ResidentialHigh), "Residential (High)");
-        assert_eq!(zone_label(ZoneType::CommercialLow), "Commercial (Low)");
-        assert_eq!(zone_label(ZoneType::CommercialHigh), "Commercial (High)");
-        assert_eq!(zone_label(ZoneType::Industrial), "Industrial");
-        assert_eq!(zone_label(ZoneType::Office), "Office");
-        assert_eq!(zone_label(ZoneType::MixedUse), "Mixed Use");
-    }
-
-    #[test]
-    fn test_education_label() {
-        assert_eq!(education_label(0), "None");
-        assert_eq!(education_label(1), "Elementary");
-        assert_eq!(education_label(2), "High School");
-        assert_eq!(education_label(3), "University");
-        assert_eq!(education_label(4), "Advanced");
-    }
-
-    #[test]
-    fn test_citizen_name_deterministic() {
-        let entity = Entity::from_raw(42);
-        let name1 = citizen_name(entity, Gender::Male);
-        let name2 = citizen_name(entity, Gender::Male);
-        assert_eq!(name1, name2, "Names should be deterministic");
-    }
-
-    #[test]
-    fn test_citizen_name_gender_difference() {
-        let entity = Entity::from_raw(0);
-        let male_name = citizen_name(entity, Gender::Male);
-        let female_name = citizen_name(entity, Gender::Female);
-        assert_ne!(male_name, female_name, "Male/female names should differ");
-    }
-
-    #[test]
-    fn test_search_state_default() {
-        let state = SearchState::default();
-        assert!(!state.visible);
-        assert!(state.query.is_empty());
-        assert!(state.search_buildings);
-        assert!(state.search_citizens);
-        assert!(state.building_results.is_empty());
-        assert!(state.citizen_results.is_empty());
     }
 }
